Clarified UI messages in tmimport.sh
Previously used the language "attributes are now" to describe the
extended attributes matching a backup drive to a computer. However,
that is both overly erudite language and completely meaningless to
the average user. Replaced it with the more meaningful language:
"This backup drive is currently matched to the following computer:" and
"This backup drive has been matched to the current computer:", and
omitted the names of the extended attributes.